feat: add context.version property
