+++ .changeset/metal-hounds-tie.md
@@ -1,0 +1,7 @@
+---
+'@rsbuild/webpack': patch
+'@rsbuild/shared': patch
+'@rsbuild/core': patch
+---
+
+feat: add context.version property
