+++ packages/core/src/rspack-provider/index.ts
@@ -2,3 +2,8 @@ export { getRspackVersion } from './shared';
 export { rspackProvider } from './provider';
 export type { RspackProvider } from './provider';
 export type { Rspack, RspackConfig } from '@rsbuild/shared';
+export {
+  createPublicContext,
+  createContextByConfig,
+  updateContextByNormalizedConfig,
+} from './core/createContext';
