+++ packages/document/docs/en/api/javascript-api/instance.mdx
@@ -6,6 +6,10 @@ This section describes all the properties and methods on the Rsbuild instance ob
 
 `rsbuild.context` is a read-only object that provides some context infos.
 
+### rsbuild.context.version
+
+The version of `@rsbuild/core` currently in use.
+
 ### rsbuild.context.entry
 
 The entry object from the `source.entry` option.
