+++ packages/document/docs/zh/api/javascript-api/instance.mdx
@@ -6,6 +6,10 @@
 
 `rsbuild.context` 是一个只读对象，提供一些上下文信息。
 
+### rsbuild.context.version
+
+当前使用的 `@rsbuild/core` 版本。
+
 ### rsbuild.context.entry
 
 构建入口对象，对应 `source.entry` 选项。
