+++ packages/shared/src/index.ts
@@ -2,7 +2,6 @@ export * from './color';
 export * from './constants';
 export * from './createHook';
 export * from './pluginStore';
-export * from './createContext';
 export * from './utils';
 export * from './fs';
 export * from './getBrowserslist';
